# openapi.yaml
...
openapi: 3.1.0
info:
  title: Silicon Hook API
  version: 0.5.0
  summary: Signed webhook endpoints with ordered, acknowledged delivery to Silicons.
  description: |
    Silicon Hook gives each Silicon signed webhook endpoints at
    `https://hook.teamofsilicons.com/silicon/{silicon_id}/{endpoint_key}`.
    Every request to an active endpoint is answered with `webhook.ok`. Requests
    that verify against the hook's signature policy are retained for 14 days
    and delivered in order over the WebSocket stream at `/api/v1/ws` or by
    polling `/api/v1/silicons/{silicon_id}/deliveries`; requests that do not
    verify are withheld and kept in a separate 14-day blocked log. Timestamps
    are UTC RFC 3339 values. Endpoint keys route requests but never
    authenticate them. IAM test app_secret selectors choose isolated sandboxes via
    X-Hook-Test-App-Secret; they never grant actor authority. Deprecated API v1
    contracts sunset only after seven request-free days. Contract accounting is
    independent of optional diagnostic telemetry.
  contact:
    email: engineering@teamofsilicons.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://backend.hook.teamofsilicons.com
  description: Backend API; public ingress may use the configured Hook frontend origin.
security:
- bearerAuth: []
...
paths:
...
  /api/v1/relay/ws:
    get:
      operationId: multiplexedRelay
      summary: Prewarm one physical socket and independently authorize subscriptions
      description: >-
        Upgrade to a WebSocket with relay protocol 1. The server sends relay_ready.
        Submit subscribe frames with a unique subscription_id, bearer token, org_id,
        silicon_ids and optional app_secret or test_key. Per-identity messages are
        wrapped as type frame with subscription_id and the ordinary v1 frame.
        Authorization failures affect only the logical subscription. The prewarm
        connection grants no access. Limit 256 subscriptions and 64 KiB client frames.
        See https://docs.hook.teamofsilicons.com/client/relay/ for heartbeats and ACKs.
      tags: [Deliveries]
      security: []
      responses:
        '101':
          description: Physical socket ready; each subscription still requires authorization
        '400':
          description: Invalid upgrade
  /api/v1/telemetry:
    post:
      operationId: recordTelemetryEvent
      summary: Append a bounded diagnostic event to Hook-owned storage
      description: >-
        Requires a live IAM actor in the selected environment. Maximum 8 KiB.
        Credentials, free-form text and payload fields are rejected. Event IDs
        deduplicate within the environment. Accepted writes may be dropped under
        load or during shutdown. Send X-Hook-Telemetry off to opt out.
      tags: [System]
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/TestAppSecret'
      - name: X-Hook-Telemetry
        in: header
        schema: {type: string, enum: [on, off]}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryEvent'
      responses:
        '202':
          description: Accepted for best-effort persistence
        '204':
          description: Collection disabled
        '401':
          description: Invalid actor or environment credentials
        '422':
          description: Invalid event shape or value
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM opaque access token
  parameters:
    TestAppSecret:
      name: X-Hook-Test-App-Secret
      in: header
      required: false
      description: >-
        IAM test application selector, validated online on each request. Mutually
        exclusive with X-Hook-Test-Key. Never send on ingress or root administration
        routes. Invalid selectors fail closed; they never select production.
      schema:
        type: string
        pattern: '^ask_[A-Za-z0-9_-]{43}$'
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        "$ref": "#/components/schemas/OrganizationId"
...
  schemas:
    TelemetryEvent:
      type: object
      additionalProperties: false
      required: [event_id, trace_id, source, step, outcome, version]
      properties:
        event_id: {type: string, format: uuid}
        trace_id: {type: string, format: uuid}
        source: {type: string, enum: [cli, daemon, client, web]}
        step: {type: string, enum: [command, connect, subscribe, deliver, ack, refresh, update, page_view, interaction, error]}
        outcome: {type: string, enum: [started, succeeded, failed, retrying, skipped]}
        version: {type: string, minLength: 1, maxLength: 32, pattern: '^[0-9.]+$'}
        operation:
          type: string
          enum: [login, logout, whoami, refresh, webhook, unhook, hooks, events, blocked, deliveries, env, config, daemon, docs, commands, system, report, about, overview, live, testing, connections, relay, iam, create, set-secret, list, show, update, delete, restore, enable, disable, rotate, connect-iam, listen]
        duration_ms: {type: integer, minimum: 0}
        progress: {type: integer, minimum: 0, maximum: 4294967295}
        status: {type: integer, minimum: 0, maximum: 65535}
        os: {type: string, enum: [linux, macos, windows, freebsd, openbsd, netbsd, android, ios, unknown]}
        arch: {type: string, enum: [x86_64, aarch64, x86, arm, riscv64, wasm32, unknown]}
    OrganizationId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^[!-~]+$"
...